link_demo : code added to demonstrate new scenario
code is added to add a node in multiple lists.
Now process node is added to four lists depending
on user choice. There are four lists available,
using, waiting, existing. if process is existing
and currently waiting to be used it should be present
in both the existing and waiting list.
If process is existing and currently being used
it should be present in both existing and using list

signed-of-by: Lokesh N Jaliminche <lokesh.jaliminche@gmail.com>